Move core quicklinks-related code inside features/quick-links directory
