+++ src/components/settings/tabs/SettingsFormQuickLinksTab.svelte
@@ -2,7 +2,7 @@
 import { createEventDispatcher } from 'svelte';
 import Switch from '@components/Switch.svelte';
 import SettingsFormQuickLinksField from '@components/settings/fields/SettingsFormQuickLinksField.svelte';
-import { frequentLinksSupported } from '@stores/frequent-links';
+import { frequentLinksSupported } from '@features/quick-links/store';
 
 export let data = {
   quickLinks: [],
